move Container to pojos package
