+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/pojo/Container.java
@@ -1,4 +1,4 @@
-package org.batfish.common;
+package org.batfish.datamodel.pojo;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
