+++ projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/pojo/ContainerTest.java
@@ -1,4 +1,4 @@
-package org.batfish.common;
+package org.batfish.datamodel.pojo;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
