+++ projects/coordinator/src/main/java/org/batfish/coordinator/WorkMgrServiceV2.java
@@ -4,7 +4,6 @@ import static org.batfish.common.CoordConsts.DEFAULT_API_KEY;
 import static org.batfish.common.CoordConsts.SVC_KEY_API_KEY;
 
 import java.util.List;
-import javax.annotation.Nullable;
 import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
 import javax.ws.rs.HeaderParam;
@@ -16,8 +15,8 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.UriInfo;
-import org.batfish.common.Container;
 import org.batfish.common.CoordConsts;
+import org.batfish.datamodel.pojo.Container;
 
 /**
  * The Work Manager is a RESTful service for servicing client API calls.
