+++ projects/coordinator/src/test/java/org/batfish/coordinator/WorkMgrServiceTest.java
@@ -9,9 +9,9 @@ import java.util.SortedSet;
 import java.util.TreeSet;
 import javax.ws.rs.core.Response;
 import org.batfish.common.BatfishLogger;
-import org.batfish.common.Container;
 import org.batfish.common.util.BatfishObjectMapper;
 import org.batfish.coordinator.config.Settings;
+import org.batfish.datamodel.pojo.Container;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
