+++ projects/coordinator/src/test/java/org/batfish/coordinator/WorkMgrTest.java
@@ -13,8 +13,8 @@ import java.util.TreeSet;
 import org.batfish.common.BatfishException;
 import org.batfish.common.BatfishLogger;
 import org.batfish.common.BfConsts;
-import org.batfish.common.Container;
 import org.batfish.coordinator.config.Settings;
+import org.batfish.datamodel.pojo.Container;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
